requirements.txt
logger==1.4
numpy==1.26.1
python-dotenv==1.0.0
